feat(init): added task boilerplate; add def stubs
This commit adds a parser interface for interfacing with ArgumentParser
methods.  There are various stubs for matching any arguments (args) that
the parser returns from `Parser(...).parse()`. There *is* basic
boilerplate functionality such as a --verbose flag that interacts with a
very simple logging configuration.

Additionally, this is being developed on Nix right now, so there is also
a shell.nix file included for various reasons.

Signed-off-by: Joshua Rose <joshuarose@gmx.com>
Authored-by: Joshua Rose <joshuarose@gmx.com>