feat(command): Added implementation for `--init`
This commit addresses the need for the implementation of the --init flag
wherein files are deleted and re-created if the command is specified.
Note that it might be of use some time in the future to create a prompt
and/or a flag such as --overwrite to ensure that the user means to
overwrite their existing task files.

This commit also addresses the need for cross-module-logging
capabilities. Each command is passed a mandatory logging object so that
it can be used for outputting information (such as the
creation/deletion/modification of files) that may be warranted.

Signed-off-by: Joshua Rose <joshuarose@gmx.com>
Authored-by: Joshua Rose <joshuarose@gmx.com>