feat(license): add `--license` command
This commit adds a license command.

The way that this happens is somewhat interesting. It should be noted
that it would be *very* rare that the LICENSE file could not be read
locally. The only reason why this would ever be the case is that the
repository was cloned in a place where files could not be read, and eve
nthen that would mean that the script wouldn't even be able to run in
the first place.

So I suppose this is just the fallback for the fallback, but in the case
that LICENSE cannot be read locally, it is actually read from online.
The internet connection is tested (by opening a socket to 8.8.8.8:53)
and then urllib is used to get the raw text data from github from the
online LICENSE file.

Signed-off-by: Joshua Rose <joshuarose@gmx.com>
Authored-by: Joshua Rose <joshuarose@gmx.com>