feat(misc): replaced `epilogue` field with CVS URL
This commit simply replaces the `epilogue` value (author) with a new value
(the GitHub URL). This is more relevant as it links to the author in the
first place, and - in addition, links to the online version of this
program which can be possibly more useful.

This might be helpful as well in the case that a copy of this program
was distributed without any CVS metadata.

Signed-off-by: Joshua Rose <joshuarose@gmx.com>
Authored-by: Joshua Rose <joshuarose@gmx.com>