dist: migrate `setup.py` -> `pyproject.toml`
This commit does a few things.

1. Migrate setup.py to pyproject.toml
2. Add additional dependencies to shell.nix
3. Modify commands.py to support --version command

Migrates setup.py to pyproject.toml: pyproject.toml is more 'modern' and
is considered 'best-practice' which is why this project is switching to
it. Additionally, the module used to get the project version prefers
pyproject.toml to be used.

Add additional dependencies to shell.nix: Add a dependency that helps
get the project version from git. Additionally, importlib-metadata
allows the version to be imported within Python.

Modify commands.py to support --version: This uses the aforementioned
modules to support the version command, printing it to the console when
called.

Resolves: #2

Signed-off-by: Joshua Rose <joshuarose@gmx.com>
Authored-by: Joshua Rose <joshuarose@gmx.com>