feat(task): added task IO
This commit addresses a component of the functionality that is needed
for the first initial release: writing tasks. The user can essentially
just write tasks to a json file.

To make it easier to read, the tasks are pretty printed. If this becomes
a problem later on, I might make a configuration file that allows the
user to specify if they want tasks to be pretty printed or not. This is
just because there might be lots and lots of tasks, which may ultimately
build up a lot of space.